Relative type names
### Relative type names
Type names are now shortened based on the scope they are output in. Although it does producer better looking code the prime reason was to reduce errors when creating nested types or intermodule types.

### Module -> Namespace
`ModuleSpec` now supports outputting `namespace` and  `module` defintions. The API remains the same but adds a `type` parameter for selecting between them.

### Constructor Argument & Automatic Constructor Argument Properties
Previously `ClassSpec` always attempted to turn properties into constructor arguments. This can cause problems with optioal arguments. This is now selectable via flag on `ClassSpec` and is enabled by defualt to match previous behavior.

The real change comes from the fact that when this option is turned on constructor parameters are not allowed to use the optional syntax (aka `?`); instead `| undefined` is appended to optional parameters.

Finally, previously `ClassSpec` would allow non-optional parameters to be generated after optional ones; which is not allowed. Through these changes this bug has been fixed.